+++ TruRating.TruModule.V2xx.Tests.Unit/Network/SystemWebClientTests/WhenCreatingAWebRequest.cs
@@ -19,23 +19,28 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-using System.Runtime.Serialization;
+
+using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using TruRating.Dto.TruService.V2xx.Http;
-using TruRating.TruModule.V2xx.Device;
 using TruRating.TruModule.V2xx.Network;
 
-namespace TruRating.TruModule.V2xx.Tests.Unit.Network
+namespace TruRating.TruModule.V2xx.Tests.Unit.Network.SystemWebClientTests
 {
     [TestClass]
-    public class WhenConstructingDefaultTruServiceClient : MsTestsContext<V2xx.Network.TruServiceHttpClient>
+    public class WhenCreatingAWebRequest : MsTestsContext
     {
-        [TestMethod]
-        public void ItShouldHaveTheDefaultSerializerAndSystemWebClient()
+        private SystemWebClient _sut;
+
+        [TestInitialize]
+        public void Setup()
         {
-            V2xx.Network.TruServiceHttpClient truServiceHttpClient = (V2xx.Network.TruServiceHttpClient)V2xx.Network.TruServiceHttpClient.CreateDefault(100, "", MockOf<ILogger>(),MockOf<V2xx.Security.IMacSignatureCalculator>());
-            Assert.IsTrue(truServiceHttpClient.Serializer.GetType() == typeof(V2xx.Serialization.DefaultSerializer));
-            Assert.IsTrue(truServiceHttpClient.WebClientFactory.GetType() == typeof(SystemWebClientFactory));
+            _sut = new SystemWebClient(1000);
+        }
+        [TestMethod]
+        public void ItShouldHaveTheTimeoutSet()
+        {
+            var webRequest = _sut.GetWebRequestTimeout(new Uri("http://localhost"));
+            Assert.IsTrue(webRequest.Timeout == 1000);
         }
     }
 }
